feat: add lints.toml, fix linter issues (#291)

+++ src-tauri/src/process_killer.rs
@@ -1,9 +1,7 @@
-use log::warn;
-use std::process::Command;
-
+#[cfg(target_os = "windows")]
 const LOG_TARGET: &str = "tari::universe::process_killer";
 
-pub fn kill_process(pid: u32) -> Result<(), anyhow::Error> {
+pub fn kill_process(pid: i32) -> Result<(), anyhow::Error> {
     #[cfg(target_os = "windows")]
     {
         let output = Command::new("taskkill")
@@ -19,7 +17,8 @@ pub fn kill_process(pid: u32) -> Result<(), anyhow::Error> {
         use nix::sys::signal::{self, Signal};
         use nix::unistd::Pid;
 
-        let pid = Pid::from_raw(pid as i32);
+        let pid = Pid::from_raw(pid);
+
         let _ = signal::kill(pid, Signal::SIGTERM);
     }
     Ok(())
